+++ crates/rexafs/src/prelude.rs
@@ -20,7 +20,7 @@ pub use crate::xafs::normalization::{Normalization, NormalizationMethod};
 #[cfg(feature = "ndarray-compat")]
 pub use crate::xafs::nshare::{ToNalgebra, ToNdarray1};
 pub use crate::xafs::xafsutils::{FTWindow, XAFSUtils};
-pub use crate::xafs::xrayfft::{FFTUtils, XrayFFTF, XrayFFTR};
+pub use crate::xafs::xrayfft::{FFTGrid, FFTUtils, XrayFFTF, XrayFFTR};
 
 #[cfg(feature = "plotting")]
 pub use crate::plot::{PlotError, PlotXAS, XASPlotBuilder};